Made changes to allow Solaris5 platform to compile with gcc-warnings enabled
git-svn-id: svn://svn.clusterresources.com/torque/branches/2.4-fixes@3393 8f940c70-5916-0410-ac46-d1fa2fa6ea29

+++ src/resmom/catch_child.c
@@ -1345,7 +1345,7 @@ static void preobit_reply(
     if (LOGLEVEL >= 2)
       {
       snprintf(log_buffer, 1024, "epilog subtask created with pid %d - substate set to JOB_SUBSTATE_OBIT - registered post_epilogue",
-               cpid);
+               (int)cpid);
 
       log_record(
         PBSEVENT_DEBUG,
